feat 회원가입 & atoms => refector strage 디렉토리 구조 개선
