+++ client/src/GlobalComponents/Banners/RecruitmentBanner.tsx
@@ -5,7 +5,8 @@ import * as S from "./SRecruitmentBanner";
 import dummyRecruitmentData from "../../assets/dummyData/dummyRecruitment.json";
 import moment from "moment";
 import { useRecoilState } from "recoil";
-import { recruitmentInfoAtom, isRecruitmentAtom } from "../../atoms";
+import { recruitmentInfoAtom } from "../../storage/common";
+import { isRecruitmentAtom } from "../../storage/common";
 import { useEffect, useState } from "react";
 import Moment from "react-moment";
