+++ client/src/GlobalComponents/Header/Header.tsx
@@ -2,7 +2,9 @@ import * as S from "./HeaderStyle";
 import * as T from "../Text/Text";
 import { Link, useNavigate } from "react-router-dom";
 import { useRecoilState, useRecoilValue } from "recoil";
-import { isAdmin, isLoggedInAtom, myUserIdAtom } from "../../atoms";
+import { isAdmin } from "../../storage/common";
+import { myUserIdAtom } from "../../storage/user";
+import { isLoggedInAtom } from "../../storage/common";
 import PersonalInfo from "../../assets/dummyData/dummyPersonalInfo.json";
 import { useEffect, useState } from "react";
 import { useAnimation, useScroll } from "framer-motion";
