+++ client/src/Types.ts
@@ -1,5 +1,4 @@
 import { SetterOrUpdater } from "recoil";
-import { IAdminFileUploadereAtom } from "./atoms";
 
 export enum Categories {
   "self_dev"="self_dev",
@@ -13,17 +12,9 @@ export enum Categories {
 //api post & put
 
 export interface IJoinBody {
-  name: string;
+  nickname: string;
   email: string;
   password: string;
-  birthYear?: number;
-  depaartment?: string;
-  phone?: string;
-  studentId?: string;
-  introduce?: string;
-  team?: string;
-  role?: string;
-  isChief?: boolean;
 };
 
 export interface ISignupBody {
@@ -37,8 +28,13 @@ export interface IPasswordResetBody {
   newPassword:string;
 };
 
-
-
+//Recruitment
+export interface IRecruitment {
+  recruitment_link?:string;
+  inquiry?:string;
+  due_at?:string;
+  content?:string;
+};
 
 //Community Page
 export interface ICategoriesData {
@@ -213,9 +209,9 @@ export interface IDraggableCardProps {
 //Archieve
 
 export interface IVideoItems {
-  id: string;
   playlistId : string;
   publishedAt : string;
+  id: string;
   title: string;
   description: string;
   thumnailUrl: string | undefined;
@@ -314,10 +310,10 @@ export interface IMemberResetPWForm {
 
 export interface IMemberSignupForm {
   extraError: string;
-  username?: string;
-  email?:string;
-  password?: string;
-  passwordConfirm?:string;
+  username: string;
+  email:string;
+  password: string;
+  passwordConfirm:string;
 };
 
 export interface IMemberSignupDetailForm {
