+++ client/src/assets/images/commons/index.d.ts
@@ -1,0 +1,16 @@
+declare module "*.jpg"{
+  const path: string;
+  export default path;
+};
+declare module "*.png"{
+  const path: string;
+  export default path;
+};
+declare module "*.jpeg"{
+  const path: string;
+  export default path;
+};
+declare module "*.gif"{
+  const path: string;
+  export default path;
+};
