+++ client/src/routes/Community/Community.tsx
@@ -9,7 +9,8 @@ import { useEffect, useState } from "react";
 import Loading from "../../GlobalComponents/Loading/Loading";
 import { Link, Outlet, useNavigate, useParams } from "react-router-dom";
 import { useSetRecoilState, useRecoilState, useRecoilValue } from "recoil";
-import { communityCategoryState, communityPageIndex, communitypagenationIndex, contentTitleAtom, isLoggedInAtom, communityCategoryTitleAtom, communityCategorySubTitleAtom } from "../../atoms";
+import { communityCategoryState, communityPageIndex, communitypagenationIndex, contentTitleAtom, communityCategoryTitleAtom, communityCategorySubTitleAtom } from "../../storage/community";
+import { isLoggedInAtom } from "../../storage/common";
 import Footer from "../../GlobalComponents/Footer/Footer";
 import { ICategories, ICategoriesData } from "../../Types";
 import { useCommunityResultHandler } from "../../hooks/useCommunityResultHandler";
