+++ client/src/routes/Home/components/IndexByCategory/IndexByCategory.tsx
@@ -3,7 +3,8 @@ import * as T from "../../../../GlobalComponents/Text/Text";
 import * as G from "../../../../GlobalComponents/Spacing/Spacing";
 import * as R from "../Recommendation/SRecommends";
 import { useRecoilValue } from "recoil";
-import { playlistItemState, isLoadingAtom } from "../../../../atoms";
+import { playlistItemState } from "../../../../storage/archive";
+import { isLoadingAtom } from "../../../../storage/common";
 import { useState } from "react";
 import Loading from "../../../../GlobalComponents/Loading/Loading";
 import VideoCard from "../Recommendation/VideoCard";
