+++ client/src/routes/Member/components/Login/Login.tsx
@@ -5,7 +5,9 @@ import SubNextBtn from "../../../../GlobalComponents/Buttons/SubNextBtn";
 import { Link, useNavigate } from "react-router-dom";
 import { useForm } from "react-hook-form";
 import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
-import { isAdmin, isLoggedInAtom, myUserIdAtom } from "../../../../atoms";
+import { isAdmin } from "../../../../storage/common";
+import { myUserIdAtom } from "../../../../storage/user";
+import { isLoggedInAtom } from "../../../../storage/common";
 import dummyLogInResponse from "../../../../assets/dummyData/dummyLogInResponse.json";
 import dummyPersonalInfo from "../../../../assets/dummyData/dummyPersonalInfo.json";
 import { useState } from "react";
