+++ client/src/storage/user.ts
@@ -1,6 +1,15 @@
 import { atom } from "recoil";
+import { v1 } from "uuid";
+import { IMemberPersonalData } from "../Types";
 
-export const usersAtom = atom({
-  key: 'users',
-  default: null
+/**userId storage */
+export const myUserIdAtom = atom<number | null>({
+  key:`userState/${v1()}`,
+  default: null,
+});
+
+/**유저 개인정보 */
+export const myPersonalDataAtom = atom<IMemberPersonalData>({
+  key:`userState/${v1()}`,
+  default: {}
 });
